refactor: split sprite editor into subdirectory with per-concern files
Move src/editor/sprite_editor.{h,cpp} into src/editor/sprite_editor/
and split the 1153-line implementation across six translation units:
sprite_editor.cpp (Draw + DrawDataEditor), sprite_editor_io.cpp
(Load/Import/Save), sprite_editor_undo.cpp, sprite_editor_preview.cpp,
sprite_editor_frames.cpp (DrawFramesPane), and sprite_editor_clips.cpp
(DrawClipsPane). Update CMakeLists.txt and editor_layer.h accordingly.
